pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.19.0",
    "aiogram-dialog>=2.3.1",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "orjson>=3.10.16",
    "pydantic-settings>=2.8.1",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
]
